pyproject.toml
[project]
name = "model-development"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.39.12",
    "loguru>=0.7.3",
    "mlflow>=3.1.4",
    "numpy>=1.26.4",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "tensorflow==2.15.0",
    "tqdm>=4.67.1",
    "typer>=0.16.0",
]
